move files to controllers
https://coveord.atlassian.net/browse/KIT-273
